provjeriti jos detalje oko user/services/reservations da vidimo tko ima autorizaciju za raditi promjene itd

+++ golfBackend/resources/reservations.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy import and_, or_, func
 from datetime import datetime, timedelta
@@ -15,49 +15,9 @@ from flask_mail import Message
 from flask import current_app
 from utils import get_all_admin_mails
 from extensions import mail
+from utils import confirmation_mail, cancelation_mail
+from utils import admin_required
 
-
-
-def confirmation_mail(to_mail,admin_mail, reservation, usluge):
-    usluge_str = ", ".join(usluge)
-
-    msg = Message(
-        subject="Potvrda rezervacije -Golf Resort",
-        sender= current_app.config["MAIL_USERNAME"],
-        recipients=[to_mail],
-        body=f"""
-Poštovani,
-
-uspješno ste rezervirali termin u Golf Resortu.
-Usluge: {usluge_str}
-Datum: {reservation.date}
-Vrijeme: {reservation.start_time.strftime("%H:%M")}- {reservation.end_time.strftime('%H:%M')}
-Trajanje: {reservation.duration_minutes} minuta
-
-Zahvaljujemo na povjerenju!
-"""
-    )
-    admin_msg = Message(
-        subject="Obavijest o novoj rezervaciji - Golf Resort",
-        sender= current_app.config["MAIL_USERNAME"],
-        recipients= admin_mail,
-        body=f"""
-        Nova rezervacija je upravo izvršena.
-Usluge: {usluge_str}
-Datum: {reservation.date}
-Vrijeme: {reservation.start_time.strftime("%H:%M")}- {reservation.end_time.strftime('%H:%M')}
-Trajanje: {reservation.duration_minutes} minuta
-
-Provjerite sustav za dodatne informacije.
-"""
-    )
-    
-    try:
-        mail.send(msg)
-        mail.send(admin_msg)
-    except Exception as e:
-        print("❌ Email sending failed:", e)
-        abort(500, message="Reservation saved, but failed to send confirmation email.")
 
 
 blp = Blueprint("Reservations", __name__, description="Operations on reservations")
@@ -65,10 +25,15 @@ blp = Blueprint("Reservations", __name__, description="Operations on reservation
 @blp.route("/reservations")
 class CreateReservation(MethodView):
     @jwt_required()
-    
     @blp.arguments(ReservationSchema(session=db.session))
     @blp.response(201, ReservationSchema)
     def post(self, reservation_data):
+        current_user_id=get_jwt_identity()
+        claims = get_jwt()
+
+        if reservation_data.user_id != current_user_id and claims.get("role") != "admin":
+            abort(403, message="Možete napraviti rezervaciju samo za svoj profil.")
+
         user_id = reservation_data.user_id
         date = reservation_data.date
         start_time = reservation_data.start_time
@@ -151,12 +116,25 @@ class CreateReservation(MethodView):
 class ReservationOperations(MethodView):
     def delete(self, reservation_id):
         reservation = ReservationModel.query.get_or_404(reservation_id)
+        if not reservation.reservation_items:
+            abort (404, message="Nisu pronađene usluge za ovu rezervaciju.")
+        user_id=reservation.user_id
+        service_names=[]
+        for item in reservation.reservation_items:
+            service_id = item.service_id
+            service = db.session.get(ServiceModel, service_id)
+            service_names.append(service.name)
         try:
+            user= db.session.get(UserModel, user_id)
+            admin_mails= get_all_admin_mails()
+            cancelation_mail(user.email, admin_mails, reservation, service_names)
             db.session.delete(reservation)
             db.session.commit()
-            return {"message": "Reservation deleted successfully."}
+            return {"message": "Rezervacija uspješno otkazana."}
         except SQLAlchemyError:
-            abort(500, message="Failed to delete reservation.")
+            abort(500, message="Došlo je do greške prilikom otkazivanja rezervacije. Pokušajte ponovno.")
+
+
     @blp.arguments(ReservationSchema(session=db.session))
     @blp.response(201, ReservationSchema)
     def put(self,updated_data, reservation_id):
@@ -231,6 +209,8 @@ from datetime import datetime
 
 @blp.route("/reservations/by-date/<string:date_str>")
 class ReservationByDate(MethodView):
+    @jwt_required()
+    @admin_required    
     @blp.response(200, ReservationSchema(many=True))
     def get(self, date_str):
         try:
@@ -263,8 +243,10 @@ class ReservationByUserId(MethodView):
         
         return reservations
 
-@blp.route("/reservations/by-category/<service_category>")
+@blp.route("/reservations/by-category/<service_category>")  
 class ReservationByCategory(MethodView):
+    @jwt_required()
+    @admin_required  
     @blp.response(200, ReservationSchema(many=True))
     def get(self, service_category):
         reservations = (
